Set up the World
Draw the world! The blue square is the "snake", the red square is the "food". The snake and the food are both 20 pixels by 20 pixels at the starting point. 

The snake should start in the top left corner of the world. The food is placed randomly as long as its x and y values are both multiplies of 20.