Milestone #2: Animate
Each time through the animation loop the snake should move player 20 pixels in the direction it is traveling. The directions are either left, right, up or down. At the start the snake should be traveling to the right

Once this milestone is completed, the program only animates the snake moving right.