Milestone #3: Handle Key Press
the direction that the snake is traveling can either be Left, Right, Up or Down and should be controlled by the keyboard. 

At any point in a graphics program we can ask the canvas for the last key press by using : key = canvas.get_last_key_press()

The key variable will then hold the name of the key which was last pressed (or None if no key has been pressed). For example, If the name of the key is "ArrowLeft" that means the last key the user pressed was the left arrow.

We process canvas.get_last_key_press() inside the animation loop. It is recommended to keep a variable which stores the current direction of travel.